Create windowstorage interface
