+++ lib/msal-browser/src/utils/BrowserConstants.ts
@@ -13,6 +13,8 @@ export const BrowserConstants = {
     CACHE_LOCATION_LOCAL: "localStorage",
     // Session storage constant string
     CACHE_LOCATION_SESSION: "sessionStorage",
+    // In memory storage string
+    CACHE_LOCATION_IN_MEMORY: "inMemoryStorage",
     // Interaction status key (only used for browsers)
     INTERACTION_STATUS_KEY: "interaction.status",
     // Interaction in progress cache value
